requirements.txt
asgiref==3.2.10
Django==3.1.2
django-taggit==1.3.0
django-widget-tweaks==1.4.8
gunicorn==20.0.4
Pillow==8.0.1
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2020.1
sorl-thumbnail==12.6.3
sqlparse==0.4.1
